Moved register-agent-service
